requirements.txt
dj-database-url==0.4.1
Django==1.9
djangorestframework==3.3.3
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
requests==2.11.1
django-extensions==1.7.4
python-dateutil==2.5.3
pytz==2016.7
